minor fix to uPlotView to suppress compile warning
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@7207 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/uPlotViewer/UPV_GUI.cpp
@@ -122,7 +122,7 @@ void UPV_GUI::addScopeVar(string varname)
   if(vectorContains(m_scope_vars, varname))
     return;
   
-  int index = (int)(m_scope_vars.size());
+  long int index = (int)(m_scope_vars.size());
   m_scope_vars.push_back(varname);
 
   string entry = "HotVar/" + varname;
